integrating the demo and main project
