--- src/main/java/uk/ac/bris/cs/hiddenmuseum/HiddenmuseumApplication.java
@@ -1,15 +1,0 @@
-package uk.ac.bris.cs.hiddenmuseum;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class HiddenmuseumApplication {
-/*
-	public static void main(String[] args) {
-		SpringApplication.run(HiddenmuseumApplication.class, args);
-	}
-
- */
-
-}
