--- src/test/java/uk/ac/bris/cs/hiddenmuseum/HiddenmuseumApplicationTests.java
@@ -1,13 +1,0 @@
-package uk.ac.bris.cs.hiddenmuseum;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class HiddenmuseumApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
